Refactorizar los fragments
Se ha eliminado código innecesario de las plantillas de los fragments.
