+++ app/src/main/java/com/igu/pepperprofesor/fragment/TitleFragment.java
@@ -18,15 +18,9 @@ public class TitleFragment extends Fragment {
 
     }
 
-    @Override
-    public void onCreate(@Nullable Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-    }
-
     @Nullable
     @Override
     public View onCreateView(@NonNull LayoutInflater inflater, @Nullable ViewGroup container, @Nullable Bundle savedInstanceState) {
-        View view = inflater.inflate(R.layout.fragment_title, container, false);
-        return view;
+        return inflater.inflate(R.layout.fragment_title, container, false);
     }
 }
