+++ app/src/main/java/com/igu/pepperprofesor/fragment/answer/CorrectAnswerFragment.java
@@ -1,4 +1,4 @@
-package com.igu.pepperprofesor.fragment;
+package com.igu.pepperprofesor.fragment.answer;
 
 import android.os.Bundle;
 import android.view.LayoutInflater;
@@ -9,8 +9,8 @@ import androidx.fragment.app.Fragment;
 
 import com.igu.pepperprofesor.R;
 
-public class SubjectSelectionFragment extends Fragment {
-    public SubjectSelectionFragment() {
+public class CorrectAnswerFragment extends Fragment {
+    public CorrectAnswerFragment() {
         // Required empty public constructor
     }
 
@@ -18,6 +18,6 @@ public class SubjectSelectionFragment extends Fragment {
     public View onCreateView(LayoutInflater inflater, ViewGroup container,
                              Bundle savedInstanceState) {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_subject_selection, container, false);
+        return inflater.inflate(R.layout.fragment_answer_correct, container, false);
     }
 }
