+++ app/src/main/res/layout/fragment_question_option3.xml
@@ -7,7 +7,7 @@
     android:layout_height="match_parent"
     android:background="@drawable/fondoazulyrojo"
     android:screenOrientation="landscape"
-    tools:context=".fragment.TwoOptionQuestionFragment">
+    tools:context=".fragment.question.TwoOptionQuestionFragment">
 
     <TextView
         android:id="@+id/enunciado3"
